Avoid GCC's "cast from pointer to integer of different size" warning.
This is a bit of a stab in the dark as I'm not sure I've got these
source files compiling correctly locally. (and the warning only
reproduces on a 32bit build anyway)

llvm-svn: 204521

+++ compiler-rt/lib/profile/InstrProfilingBuffer.c
@@ -41,8 +41,8 @@ int __llvm_profile_write_buffer(char *Buffer) {
     DataSize,
     CountersSize,
     NamesSize,
-    (uint64_t)CountersBegin,
-    (uint64_t)NamesBegin
+    (uintptr_t)CountersBegin,
+    (uintptr_t)NamesBegin
   };
 
   /* Write the data. */
